requirements.txt
psycopg2-binary
pandas
ipython-sql
ipykernel
